backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
openai
pdfplumber
PyPDF2
pdfminer.six
pdf2image
pytesseract
pillow
